refactor(ui): Split pages into separate files
